async generator

+++ 0x02-python_async_comprehension/0-async_generator.py
@@ -5,7 +5,7 @@ import random
 from typing import Iterator
 
 
-async def async_generator() -> Iterator[int]:
+async def async_generator() -> Iterator[float]:
     """generator coroutine"""
     for _ in range(10):
         await asyncio.sleep(1)
